+++ packages/webapp/webapp_server.js
@@ -5,12 +5,11 @@ import { userInfo } from 'os';
 import { join as pathJoin, dirname as pathDirname } from 'path';
 import { parse as parseUrl } from 'url';
 import { createHash } from 'crypto';
-import { connect } from './connect.js';
+import express from 'express';
 import compress from 'compression';
 import cookieParser from 'cookie-parser';
 import qs from 'qs';
 import parseRequest from 'parseurl';
-import basicAuth from 'basic-auth-connect';
 import { lookup as lookupUserAgent } from 'useragent';
 import { isModern } from 'meteor/modern-browsers';
 import send from 'send';
@@ -29,13 +28,20 @@ export const WebAppInternals = {};
 
 const hasOwn = Object.prototype.hasOwnProperty;
 
-// backwards compat to 2.0 of connect
-connect.basicAuth = basicAuth;
+const createExpressApp = () => {
+  const app = express();
+  app.set('x-powered-by', false);
+  app.set('etag', false);
+  app.set('query parser', qs.parse);
+  return app;
+};
+
+WebApp.express = express;
 
 WebAppInternals.NpmModules = {
-  connect: {
-    version: Npm.require('connect/package.json').version,
-    module: connect,
+  express: {
+    version: Npm.require('express/package.json').version,
+    module: express,
   },
 };
 
@@ -1040,12 +1046,12 @@ function runWebAppServer() {
   WebAppInternals.reloadClientPrograms();
 
   // webserver
-  var app = connect();
+  var app = createExpressApp();
 
   // Packages and apps can add handlers that run before any other Meteor
-  // handlers via WebApp.rawConnectHandlers.
-  var rawConnectHandlers = connect();
-  app.use(rawConnectHandlers);
+  // handlers via WebApp.rawHandlers.
+  var rawExpressHandlers = createExpressApp();
+  app.use(rawExpressHandlers);
 
   // Auto-compress any json, javascript, or text.
   app.use(compress({ filter: shouldCompress }));
@@ -1063,16 +1069,6 @@ function runWebAppServer() {
     res.writeHead(400);
     res.write('Not a proxy');
     res.end();
-  });
-
-  // Parse the query string into res.query. Used by oauth_server, but it's
-  // generally pretty handy..
-  //
-  // Do this before the next middleware destroys req.url if a path prefix
-  // is set to close #10111.
-  app.use(function(request, response, next) {
-    request.query = qs.parse(parseUrl(request.url).query);
-    next();
   });
 
   function getPathParts(path) {
@@ -1133,13 +1129,13 @@ function runWebAppServer() {
 
   // Core Meteor packages like dynamic-import can add handlers before
   // other handlers added by package and application code.
-  app.use((WebAppInternals.meteorInternalHandlers = connect()));
+  app.use((WebAppInternals.meteorInternalHandlers = createExpressApp()));
 
   /**
-   * @name connectHandlersCallback(req, res, next)
+   * @name handlersCallback(req, res, next)
    * @locus Server
    * @isprototype true
-   * @summary callback handler for `WebApp.connectHandlers`
+   * @summary callback handler for `WebApp.handlers`
    * @param {Object} req
    * a Node.js
    * [IncomingMessage](https://nodejs.org/api/http.html#class-httpincomingmessage)
@@ -1157,7 +1153,7 @@ function runWebAppServer() {
    */
 
   /**
-   * @method connectHandlers
+   * @method handlers
    * @memberof WebApp
    * @locus Server
    * @summary Register a handler for all HTTP requests.
@@ -1167,18 +1163,18 @@ function runWebAppServer() {
    *
    * For example, `/hello` will match `/hello/world` and
    * `/hello.world`, but not `/hello_world`.
-   * @param {connectHandlersCallback} handler
+   * @param {handlersCallback} handler
    * A handler function that will be called on HTTP requests.
-   * See `connectHandlersCallback`
+   * See `handlersCallback`
    *
    */
-  // Packages and apps can add handlers to this via WebApp.connectHandlers.
+  // Packages and apps can add handlers to this via WebApp.handlers.
   // They are inserted before our default handler.
-  var packageAndAppHandlers = connect();
+  var packageAndAppHandlers = createExpressApp();
   app.use(packageAndAppHandlers);
 
   var suppressConnectErrors = false;
-  // connect knows it is an error handler because it has 4 arguments instead of
+  // express knows it is an error handler because it has 4 arguments instead of
   // 3. go figure.  (It is not smart enough to find such a thing if it's hidden
   // inside packageAndAppHandlers.)
   app.use(function(err, req, res, next) {
@@ -1341,12 +1337,20 @@ function runWebAppServer() {
 
   // start up app
   _.extend(WebApp, {
+    // New 3.x-compatible APIs
+    handlers: packageAndAppHandlers,
+    rawHandlers: rawExpressHandlers,
+    expressApp: app,
+    // Existing 2.x APIs (kept as aliases for backwards compatibility)
     connectHandlers: packageAndAppHandlers,
-    rawConnectHandlers: rawConnectHandlers,
-    httpServer: httpServer,
+    rawConnectHandlers: rawExpressHandlers,
     connectApp: app,
+    httpServer: httpServer,
     // For testing.
     suppressConnectErrors: function() {
+      suppressConnectErrors = true;
+    },
+    _suppressExpressErrors: function() {
       suppressConnectErrors = true;
     },
     onListening: function(f) {